Upraveny jednotlive moduly. Pridan modul pole, ktery obsahuje vsechny funkce, ktere upravuji pole. Upraveny importy
